Change VoiceBroadcastRecorder as Interface
